refactor(billing): render usage with formatUsageLine
